requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiopg==1.3.1
aiosqlite==0.17.0
animec==0.3.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
discord-components==1.1.4
discord-ext-games==1.1.6
discord-py-slash-command==2.3.2
discord-together==1.1.1
discord.py==1.7.3
discordSuperUtils==0.2.2
DiscordUtils==1.3.4
ffmpeg==1.4
giphy-client==1.0.0
google-api-core==1.31.0
google-api-python-client==2.11.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
motor==2.5.1
multidict==5.1.0
packaging==21.0
Pillow==8.3.1
pistonapi==1.0.1
protobuf==3.17.3
psycopg2-binary==2.9.1
py-cord==1.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyLyrics==1.1.0
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
spotipy==2.19.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
watchdog==2.1.3
yarl==1.6.3
youtube-dl==2021.6.6
